+++ moor/test/transactions_test.dart
@@ -157,7 +157,7 @@ void main() {
     // After the transaction completes, the queries should be updated
     verify(
       streamQueries.handleTableUpdates(
-          {TableUpdate.fromTable(db.users, kind: UpdateKind.update)}),
+          {TableUpdate.onTable(db.users, kind: UpdateKind.update)}),
     ).called(1);
     verify(executor.transactions.send());
   });